requirements.txt
requests
pycryptodome
python-dotenv

